+++ frontend/src/components/EditorComponent.jsx
